Web_Prototype_on_Heroku_and_Vercel/physics_web_app_nextjs_heroku/backend/physics_web_app_3/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.10
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
